+++ proto/MonadIO/Paths.hs
@@ -1,0 +1,6 @@
+module MonadIO.Paths where
+
+import FPath.AbsFile  ( AbsFile, absfile )
+
+grep :: AbsFile
+grep = [absfile|__gnugrep__/bin/grep|]